backend/requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
psycopg2-binary
PyJWT
gunicorn
requests
Pillow
pillow-heif
